Fix error when changing checkboxes
Before, it errored if you opened a file and then tried to change a
checkbox. This was because checkboxStates was of size 0, but it needs to
map to config.backups. I fixed it by resetting checkboxStates adding
config.backups.size() "false" elements to it when a config file is
loaded